wrap slider and group plan component
